agent/requirements.txt
bedrock-agentcore>=0.1.0
strands-agents>=1.14.0
strands-agents-tools
mcp
boto3
requests